refactor: move `layouts` to their own module
