+++ tinyticker/layouts/__init__.py
@@ -1,0 +1,6 @@
+from .big_price import big_price
+from .default import default
+from .logo import logo
+from .register import LAYOUTS
+
+__all__ = ["big_price", "default", "logo", "LAYOUTS"]
